+++ app/forms/adv_example_form.py
@@ -1,0 +1,8 @@
+from wtforms import RadioField
+from wtforms.validators import DataRequired
+
+from app.forms.shared_forms import ExperimentForm
+
+
+class AdvExampleForm(ExperimentForm):
+    ...
